% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plantsFINALforPackage.R
\name{run.plant.ecosystem}
\alias{run.plant.ecosystem}
\title{run.plant.ecosystem: makes an array showing changes in the plant ecosystem (location and abundance) through time}
\usage{
run.plant.ecosystem(terrain, setup.plants, numb.plants.per.sp, timesteps)
}
\arguments{
\item{terrain}{A matrix generated using the terrain function into which plants can be placed}

\item{setup.plants}{A list generated from the setup.plants function using the survival, reproduction, and competition probabilities for each plant wanted in the ecosystem}

\item{numb.plants.per.sp}{A vector of length n storing the values for the initial population size desired for each of n plant species.}

\item{timesteps}{A numeric value representing the number of timesteps over which the user wishes to simulate their ecosystem.}
}
\description{
run.plant.ecosystem: makes an array showing changes in the plant ecosystem (location and abundance) through time
}
